Many Core Updates
Compressed code down by number of lines for adding and updating visual
objects.  Made it so many of the settings can be easily set /
customized, and other values automatically calculated, without need to
recode a section or two every time the scale is adjusted.

also fixed the line-wrapping issue of showing stuff on the line before /
after when too near the edge.  this became more of an issue when i
scaled up the grid, so I went ahead and put MANY hours into fixing that.